Support assistant tool with file id
