+++ src/promptflow/promptflow/executor/_docstring_parser.py
@@ -13,12 +13,12 @@ class DocstringParser:
             field_body = field[1].astext()
 
             if field_name.startswith("param"):
-                param_name = field_name.split(' ')[1]
+                param_name = field_name.split(" ")[1]
                 if param_name not in params:
                     params[param_name] = {}
                 params[param_name]["description"] = field_body
             if field_name.startswith("type"):
-                param_name = field_name.split(' ')[1]
+                param_name = field_name.split(" ")[1]
                 if param_name not in params:
                     params[param_name] = {}
                 params[param_name]["type"] = field_body
