+++ src/promptflow/tests/test_configs/flows/assistant_tool_with_file_id/flow.dag.yaml
@@ -1,0 +1,45 @@
+inputs:
+  content:
+    type: list
+    default:
+    - type: text
+      text: Attached file contains the stock price of company A and B. Please draw a line chart with the stock price of the company A, B and C and return a CVS file with the data.
+    - type: file_path
+      file_path:
+        path: ./stock_price.csv
+  assistant_id:
+    type: string
+    default: asst_yWhdFYoCS1UatnRRQZGY85aL
+outputs:
+  assistant_output:
+    type: string
+    reference: ${assistant_play.output}
+nodes:
+- name: create_thread
+  type: python
+  source:
+    type: code
+    path: create_thread.py
+  inputs:
+    conn: openai_connection
+- name: assistant_play
+  type: python
+  source:
+    type: code
+    path: oai_assistant.py
+  inputs:
+    conn: openai_connection
+    content: ${inputs.content}
+    assistant_id: ${inputs.assistant_id}
+    thread_id: ${create_thread.output}
+    assistant_definition:
+      module: gpt-4-1106-preview
+      instructions: You are a helpful assistant.
+      tools:
+      - type: code_interpreter
+      - type: function
+        tool_type: python
+        source:
+          type: code
+          path: get_stock_price.py
+version: 2
