+++ src/promptflow/tests/test_configs/flows/assistant_tool_with_file_id/get_stock_price.py
@@ -1,0 +1,9 @@
+from promptflow import tool
+
+
+@tool
+def get_stock_price():
+    """Get the stock price for company C.
+    """
+
+    return [100, 105, 110, 115, 120, 125]
