+++ src/promptflow/tests/test_configs/flows/assistant_tool_with_file_id/oai_assistant.py
