+++ src/promptflow/tests/test_configs/flows/assistant_tool_with_file_id/stock_price.csv
@@ -1,0 +1,7 @@
+A,B
+100,110
+102,114
+101,120
+103,130
+99,140
+110,145
